Add delete dialog in admin part and change UI on quiz result

+++ app/src/main/java/com/example/flutterhub_jetpackcompose/screen/components/card/AssessmentCard.kt
@@ -15,6 +15,7 @@ import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Comment
 import androidx.compose.material.icons.filled.NotInterested
 import androidx.compose.material.icons.filled.Pending
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
@@ -23,7 +24,12 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -60,6 +66,8 @@ fun AssessmentCard(
     // Check if the userLink exists and if it's checked
     val isChecked = userLink?.checked ?: null // Default to false if no link exists for the user
     val theComment = userLink?.comment ?: null
+
+    var openDeleteDialog by remember { mutableStateOf(false) }
 
     Card(
         onClick = {
@@ -206,27 +214,49 @@ fun AssessmentCard(
 
                             // Delete button
                             Button(
-                                onClick = {
-                                    viewModel.deleteAssessment(
-                                        userId,
-                                        assessmentModel.id,
-                                        onSuccess = {
-                                            Toast.makeText(
-                                                context,
-                                                "Successfully deleted",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        },
-                                        onFailure = { error ->
-                                            Toast.makeText(context, error, Toast.LENGTH_SHORT)
-                                                .show()
-                                        }
-                                    )
-                                },
-                                colors = ButtonDefaults.buttonColors(containerColor = deleteColor),
-                                modifier = Modifier.padding(8.dp)
+                                onClick = { openDeleteDialog = true },
+                                colors = ButtonDefaults.buttonColors(containerColor = Color.Red)
                             ) {
                                 Text("Delete", color = Color.White)
+                            }
+
+                            if (openDeleteDialog) {
+                                AlertDialog(
+                                    onDismissRequest = { openDeleteDialog = false },
+                                    title = { Text("Confirm Deletion") },
+                                    text = { Text("Are you sure you want to delete this assessment?") },
+                                    confirmButton = {
+                                        TextButton(onClick = {
+                                            viewModel.deleteAssessment(
+                                                userId,
+                                                assessmentModel.id,
+                                                onSuccess = {
+                                                    Toast.makeText(
+                                                        context,
+                                                        "Successfully deleted",
+                                                        Toast.LENGTH_SHORT
+                                                    ).show()
+                                                },
+                                                onFailure = { error ->
+                                                    Toast.makeText(
+                                                        context,
+                                                        error,
+                                                        Toast.LENGTH_SHORT
+                                                    ).show()
+                                                }
+                                            )
+                                            openDeleteDialog =
+                                                false  // Close dialog after confirmation
+                                        }) {
+                                            Text("Yes")
+                                        }
+                                    },
+                                    dismissButton = {
+                                        TextButton(onClick = { openDeleteDialog = false }) {
+                                            Text("Cancel")
+                                        }
+                                    }
+                                )
                             }
                         }
                     }
