+++ app/src/main/java/com/example/flutterhub_jetpackcompose/screen/components/card/QuizCard.kt
@@ -9,13 +9,19 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -43,6 +49,7 @@ fun QuizCard(
     val editColor = if (isDarkMode) EditGreenDark else EditGreenLight
     val deleteColor = if (isDarkMode) DeleteRedDark else DeleteRedLight
 
+    var openDeleteDialog by remember { mutableStateOf(false) }
 
     Card(
         shape = RoundedCornerShape(15.dp), // Rounded corners for the card
@@ -85,26 +92,44 @@ fun QuizCard(
 
                 // Delete Button
                 Button(
-                    onClick = {
-                        viewModel.deleteQuiz(quiz.id,
-                            onSuccess = {
-                                viewModel.quizzes
-
-                                Toast.makeText(
-                                    context,
-                                    "Quiz Successfully deleted",
-                                    Toast.LENGTH_SHORT
-                                ).show()
-
-                            },
-                            onFailure = { error ->
-                                Toast.makeText(context, error, Toast.LENGTH_SHORT).show()
-                            })
-                    },
-                    colors = ButtonDefaults.buttonColors(containerColor = deleteColor)
+                    onClick = { openDeleteDialog = true },
+                    colors = ButtonDefaults.buttonColors(containerColor = Color.Red)
                 ) {
                     Text("Delete", color = Color.White)
                 }
+
+                if (openDeleteDialog) {
+                    AlertDialog(
+                        onDismissRequest = { openDeleteDialog = false },
+                        title = { Text("Confirm Deletion") },
+                        text = { Text("Are you sure you want to delete this quiz?") },
+                        confirmButton = {
+                            TextButton(onClick = {
+                                viewModel.deleteQuiz(quiz.id,
+                                    onSuccess = {
+                                        Toast.makeText(
+                                            context,
+                                            "Successfully deleted",
+                                            Toast.LENGTH_SHORT
+                                        ).show()
+                                    },
+                                    onFailure = { error ->
+                                        Toast.makeText(context, error, Toast.LENGTH_SHORT).show()
+                                    }
+                                )
+                                openDeleteDialog = false  // Close dialog after confirmation
+                            }) {
+                                Text("Yes")
+                            }
+                        },
+                        dismissButton = {
+                            TextButton(onClick = { openDeleteDialog = false }) {
+                                Text("Cancel")
+                            }
+                        }
+                    )
+                }
+
             }
         }
     }
